+++ tests/src/test/scala/cats/collections/AvlSetSpec.scala
